Use esbuild and lorca
esbuild instead of rollup.
esbuild is significantly faster, and it is written in go

lorca instead of wails
lorca is a slightly different model than wails. lorca uses the existing installed chrome version and controls it via the devtools protocol. This allows for lightweight programs which support latest web technolgies